loosing internet soon hol up

+++ src/default-modules/core/commands/module/remove.ts
@@ -1,10 +1,8 @@
 import { Arg, ArgValidators, CommandCreateInfo, tuple } from "../../../..";
-import { removeDirectory } from "../../../../lib/util";
+import { removeDirectory, existsDirectory } from "../../../../lib/util";
 const { WORD } = ArgValidators;
 
-const args = tuple([
-  new Arg("module", new WORD())
-]);
+const args = tuple([new Arg("module", new WORD())]);
 
 const createInfo: CommandCreateInfo = {
   args,
@@ -12,8 +10,14 @@ const createInfo: CommandCreateInfo = {
   async run(context) {
     const moduleName = context.getArg(args[0]);
     const modulesDir = `${this.bot.directory}/modules`;
-    removeDirectory(`${modulesDir}/${moduleName}`);
-    context.say(`:white_check_mark: Successfully deleted module '${moduleName}'`)
+    if (existsDirectory(`${modulesDir}/${moduleName}`)) {
+      removeDirectory(`${modulesDir}/${moduleName}`);
+      context.say(
+        `:white_check_mark: Successfully deleted module '${moduleName}'`
+      );
+    } else {
+      context.say(`:x: There is no module named ${moduleName}`);
+    }
   }
 };
